convert PrintSimulation listener to extension
